Move Statement to External rather than Internal
There was nothing that I wanted to hide away in internal, so I just
moved the module.
