+++ example/Indentation.hs
@@ -6,8 +6,8 @@ import Control.Lens.Plated (transform)
 import GHC.Natural (Natural)
 
 import Language.Python.Optics
-import Language.Python.Internal.Syntax (Statement)
-import Language.Python.Syntax.Whitespace
+import Language.Python.Syntax.Statement (Statement)
+import Language.Python.Syntax.Whitespace (Whitespace (Space, Tab))
 
 indentSpaces :: Natural -> Statement '[] a -> Statement '[] a
 indentSpaces n = transform (_Indent .~ replicate (fromIntegral n) Space)
