+++ example/Main.hs
@@ -9,7 +9,7 @@ import OptimizeTailRecursion
 import Indentation
 
 import Language.Python.Render (showModule)
-import Language.Python.Internal.Syntax (_Statements)
+import Language.Python.Syntax.Statement (_Statements)
 
 import qualified Data.Text.IO as StrictText
