+++ hpython-test/src/Generators/Correct.hs
@@ -31,6 +31,7 @@ import GHC.Stack
 import Language.Python.Syntax.Types (spType)
 import Language.Python.Optics
 import Language.Python.Internal.Syntax
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Whitespace
 
 import Generators.Common
