+++ hpython-test/test/Main.hs
@@ -20,6 +20,7 @@ import Language.Python.Internal.Syntax
 import Language.Python.Optics.Validated (unvalidated)
 import Language.Python.Parse (parseStatement, parseExpr, parseExprList)
 import Language.Python.Render
+import Language.Python.Syntax.Statement
 import Language.Python.Validate
 
 import Hedgehog
