+++ hpython.cabal
@@ -84,7 +84,6 @@ library
                      , Language.Python.Internal.Syntax.Operator.Binary
                      , Language.Python.Internal.Syntax.Operator.Unary
                      , Language.Python.Internal.Syntax.Numbers
-                     , Language.Python.Internal.Syntax.Statement
                      , Language.Python.Internal.Syntax.Strings
                      , Language.Python.Optics
                      , Language.Python.Optics.Validated
@@ -92,6 +91,7 @@ library
                      , Language.Python.Render
                      , Language.Python.Syntax
                      , Language.Python.Syntax.Expr
+                     , Language.Python.Syntax.Statement
                      , Language.Python.Syntax.Types
                      , Language.Python.Syntax.Whitespace
                      , Language.Python.Validate
