+++ src/Language/Python/DSL.hs
@@ -358,7 +358,14 @@ import Data.Maybe (fromMaybe)
 import Data.Semigroup ((<>))
 
 import Language.Python.Optics
-import Language.Python.Internal.Syntax hiding (Fundef, While, Call)
+import Language.Python.Internal.Syntax.AugAssign
+import Language.Python.Internal.Syntax.CommaSep
+import Language.Python.Internal.Syntax.Ident
+import Language.Python.Internal.Syntax.Strings
+import Language.Python.Internal.Syntax.Operator.Binary
+import Language.Python.Internal.Syntax.Operator.Unary
+import Language.Python.Syntax.Expr
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Types
 import Language.Python.Syntax.Whitespace
