+++ src/Language/Python/Internal/Render.hs
@@ -50,7 +50,19 @@ import qualified Data.Text as Text
 import qualified Data.Text.Lazy as Lazy
 import qualified Data.Text.Lazy.Builder as Builder
 
-import Language.Python.Internal.Syntax
+import Language.Python.Syntax.Expr
+import Language.Python.Syntax.Statement
+import Language.Python.Internal.Syntax.AugAssign
+import Language.Python.Internal.Syntax.Comment
+import Language.Python.Internal.Syntax.CommaSep
+import Language.Python.Internal.Syntax.Ident
+import Language.Python.Internal.Syntax.Import
+import Language.Python.Internal.Syntax.Module
+import Language.Python.Internal.Syntax.ModuleNames
+import Language.Python.Internal.Syntax.Numbers
+import Language.Python.Internal.Syntax.Operator.Binary
+import Language.Python.Internal.Syntax.Operator.Unary
+import Language.Python.Internal.Syntax.Strings
 import Language.Python.Internal.Render.Correction
 import Language.Python.Internal.Token (PyToken(..))
 import Language.Python.Syntax.Whitespace
