+++ src/Language/Python/Internal/Syntax.hs
@@ -22,7 +22,6 @@ module Language.Python.Internal.Syntax
   , module Language.Python.Internal.Syntax.Numbers
   , module Language.Python.Internal.Syntax.Operator.Binary
   , module Language.Python.Internal.Syntax.Operator.Unary
-  , module Language.Python.Internal.Syntax.Statement
   , module Language.Python.Internal.Syntax.Strings
   )
 where
@@ -44,7 +43,6 @@ import Language.Python.Internal.Syntax.ModuleNames
 import Language.Python.Internal.Syntax.Numbers
 import Language.Python.Internal.Syntax.Operator.Binary
 import Language.Python.Internal.Syntax.Operator.Unary
-import Language.Python.Internal.Syntax.Statement
 import Language.Python.Internal.Syntax.Strings
 
 reservedWords :: [String]
