+++ src/Language/Python/Internal/Syntax/IR.hs
@@ -41,7 +41,9 @@ import Language.Python.Internal.Syntax.Operator.Unary
 import Language.Python.Internal.Syntax.Strings
 import Language.Python.Syntax.Whitespace
 
-import qualified Language.Python.Internal.Syntax as Syntax
+import qualified Language.Python.Internal.Syntax.Module as Syntax
+import qualified Language.Python.Syntax.Expr as Syntax
+import qualified Language.Python.Syntax.Statement as Syntax
 
 data IRError a = InvalidUnpacking a
   deriving (Eq, Show)
