+++ src/Language/Python/Internal/Syntax/Module.hs
@@ -11,7 +11,7 @@ Portability : non-portable
 
 module Language.Python.Internal.Syntax.Module where
 
-import Language.Python.Internal.Syntax.Statement
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Whitespace
 
 -- | A Python 'Module', which is stored as a sequence of statements.
