+++ src/Language/Python/Optics.hs
@@ -24,7 +24,10 @@ import Data.Coerce (coerce)
 import Data.Function ((&))
 
 import Language.Python.Optics.Validated (unvalidated)
-import Language.Python.Internal.Syntax
+import Language.Python.Syntax.Expr (Expr (..), TupleItem (TupleUnpack), ListItem (ListUnpack), Param (..), _Exprs)
+import Language.Python.Internal.Syntax.Module
+import Language.Python.Internal.Syntax.Ident
+import Language.Python.Syntax.Statement (Block (..), CompoundStatement (..), Decorator (..), ExceptAs (..), SimpleStatement (..), Statement (..), Suite (..), _Statements)
 import Language.Python.Syntax.Types
 import Language.Python.Syntax.Whitespace
