+++ src/Language/Python/Syntax.hs
@@ -9,12 +9,13 @@ Portability : non-portable
 
 module Language.Python.Syntax
   ( module Language.Python.Syntax.Expr
+  , module Language.Python.Syntax.Statement
   , module Language.Python.Syntax.Types
   , module Language.Python.Syntax.Whitespace
   )
 where
 
 import Language.Python.Syntax.Expr
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Types
 import Language.Python.Syntax.Whitespace
-
