+++ src/Language/Python/Syntax/Types.hs
@@ -16,7 +16,10 @@ module Language.Python.Syntax.Types where
 import Control.Lens.TH (makeLenses)
 import Data.List.NonEmpty (NonEmpty)
 
-import Language.Python.Internal.Syntax hiding (Fundef, While)
+import Language.Python.Internal.Syntax.CommaSep (CommaSep, CommaSep1, CommaSep1')
+import Language.Python.Internal.Syntax.Ident (Ident)
+import Language.Python.Syntax.Expr (Arg, Expr, ListItem, Param, TupleItem)
+import Language.Python.Syntax.Statement (Decorator, ExceptAs, Suite, WithItem)
 import Language.Python.Syntax.Whitespace
 
 data Fundef v a
