+++ src/Language/Python/Validate/Indentation.hs
@@ -52,7 +52,10 @@ import qualified Data.List.NonEmpty as NonEmpty
 
 import Language.Python.Optics
 import Language.Python.Optics.Validated (unvalidated)
-import Language.Python.Internal.Syntax
+import Language.Python.Internal.Syntax.CommaSep
+import Language.Python.Internal.Syntax.Module
+import Language.Python.Syntax.Expr
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Whitespace
 import Language.Python.Validate.Indentation.Error
