+++ src/Language/Python/Validate/Syntax.hs
@@ -85,7 +85,14 @@ import qualified Data.List.NonEmpty as NonEmpty
 
 import Language.Python.Optics
 import Language.Python.Optics.Validated (unvalidated)
-import Language.Python.Internal.Syntax
+import Language.Python.Internal.Syntax (reservedWords)
+import Language.Python.Internal.Syntax.CommaSep
+import Language.Python.Internal.Syntax.Ident
+import Language.Python.Internal.Syntax.Import
+import Language.Python.Internal.Syntax.Module
+import Language.Python.Internal.Syntax.Strings
+import Language.Python.Syntax.Expr
+import Language.Python.Syntax.Statement
 import Language.Python.Syntax.Whitespace
 import Language.Python.Validate.Indentation
 import Language.Python.Validate.Syntax.Error
